chore: refactor transfer seeder
