+++ test/seed/ledger/transfer.ts
@@ -1,0 +1,63 @@
+/**
+ *  SudoSOS back-end API service.
+ *  Copyright (C) 2024  Study association GEWIS
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as published
+ *  by the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+import WithManager from '../../../src/with-manager';
+import User from '../../../src/entity/user/user';
+import Transfer from '../../../src/entity/transactions/transfer';
+import { getRandomDate } from '../helpers';
+import dinero from 'dinero.js';
+
+export default class TransferSeeder extends WithManager {
+  public async seedTransfers(
+    users: User[],
+    startDate?: Date,
+    endDate?: Date,
+  ) : Promise<Transfer[]> {
+    const transfers: Transfer[] = [];
+    const promises: Promise<any>[] = [];
+
+    for (let i = 0; i < users.length; i += 1) {
+      let date = new Date();
+      if (startDate && endDate) {
+        date = getRandomDate(startDate, endDate, i);
+      }
+      let newTransfer = Object.assign(new Transfer(), {
+        description: '',
+        amountInclVat: dinero({ amount: 100 * (i + 1) }),
+        from: undefined,
+        to: users[i],
+        createdAt: date,
+      });
+      transfers.push(newTransfer);
+      promises.push(Transfer.save(newTransfer));
+
+      newTransfer = Object.assign(new Transfer(), {
+        description: '',
+        amountInclVat: dinero({ amount: 50 * (i + 1) }),
+        from: users[i],
+        to: undefined,
+        createdAt: date,
+      });
+      transfers.push(newTransfer);
+      promises.push(Transfer.save(newTransfer));
+    }
+
+    await Promise.all(promises);
+
+    return transfers;
+  }
+}
